requirements.txt
flask
numpy
pandas
google.generativeai